pyproject.toml
[tool.poetry]
name = "llm-in-cb"
version = "0.1.0"
description = "llm project for cb"
authors = ["Egor Tishchenko & Nikita Arzamazov"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
qdrant-client = "^1.12.1"
pydantic = "^2.10.1"
torch = "^2.5.1"
fastapi = "^0.115.5"
transformers = "^4.46.3"
uvicorn = "^0.32.1"
#sentence-transformers = "^3.3.1"
asyncio = "^3.4.3"
requests = "^2.32.3"
python-telegram-bot = "^21.7"
langchain = "^0.3.9"
snscrape = "^0.7.0.20230622"
tavily-python = "^0.5.0"
langgraph = "^0.2.60"
langchain-community = "^0.3.14"
mistralai = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
